Linux build fix.
BUG=none
TEST=none
TBR=willchan

Review URL: http://codereview.chromium.org/5671001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@68609 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/spdy/spdy_network_transaction_unittest.cc
@@ -5332,7 +5332,7 @@ TEST_P(SpdyNetworkTransactionTest, ServerPushCrossOriginCorrectness) {
     'p', 'u', 's', 'h', 'e', 'd'                                 // "pushed"
   };
 
-  for (int index = 0; index < arraysize(kTestCases); index += 2) {
+  for (size_t index = 0; index < arraysize(kTestCases); index += 2) {
     const char* url_to_fetch = kTestCases[index];
     const char* url_to_push = kTestCases[index + 1];
